requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.9
importlib-metadata==1.6.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
more-itertools==8.2.0
openapi-codec==1.3.2
packaging==20.3
pluggy==0.13.1
py==1.8.1
PyJWT==2.0.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
